+++ tests/Unit/Domain.LinnApps.Tests/DailyEuReportsServiceTests/ContextBase.cs
@@ -3,7 +3,7 @@ namespace Linn.Stores2.Domain.LinnApps.Tests.DailyEuReportsServiceTests
     using Linn.Common.Persistence;
     using Linn.Common.Reporting.Models;
     using Linn.Stores2.Domain.LinnApps.External;
-    using Linn.Stores2.Domain.LinnApps.Logistics;
+    using Linn.Stores2.Domain.LinnApps.Imports;
     using Linn.Stores2.Domain.LinnApps.Reports;
 
     using NSubstitute;
